refactor(#792): split adapter-gui lib_tests into base + recent/tooltip/index suite (sweep)
lib_tests.rs 1140 -> 596, + lib_recent_tests.rs ~560 (both under cap; off debt).
The recent-projects / tooltip / block-index tests move to the sibling; the shared
chain fixtures (delay_model_ids/test_chain/input_kind/output_kind/effect_kind)
become pub(super), imported via super::tests. adapter-gui lib tests (540) pass,
zero warnings.
