+++ scripts/validate.sh
@@ -51,7 +51,6 @@ crates/adapter-gui/src/block_parameter_wiring.rs
 crates/adapter-gui/src/chain_row_wiring.rs
 crates/adapter-gui/src/compact_chain_block_handlers.rs
 crates/adapter-gui/src/desktop_app.rs
-crates/adapter-gui/src/lib_tests.rs
 crates/application/src/command.rs
 crates/application/src/local_dispatcher_tests.rs
 crates/engine/src/audio_alloc_invariant_tests.rs
